fix deployment
