+++ tests/integration/sync/conftest.py
@@ -11,6 +11,7 @@ from syftbox.client.client2 import SyftClientContext
 from syftbox.lib.client_config import SyftClientConfig
 from syftbox.lib.datasite import create_datasite
 from syftbox.lib.workspace import SyftWorkspace
+from syftbox.server.migrations import run_migrations
 from syftbox.server.server import app as server_app
 from syftbox.server.server import lifespan as server_lifespan
 from syftbox.server.settings import ServerSettings
@@ -59,9 +60,10 @@ def server_app_with_lifespan(tmp_path: Path) -> FastAPI:
     path.mkdir()
     settings = ServerSettings.from_data_folder(path)
     settings.auth_enabled = False
+    settings.otel_enabled = False
     lifespan_with_settings = partial(server_lifespan, settings=settings)
     server_app.router.lifespan_context = lifespan_with_settings
-
+    run_migrations(settings)
     return server_app
