+++ tests/unit/server/conftest.py
@@ -7,6 +7,7 @@ from fastapi.testclient import TestClient
 from syftbox.client.plugins.sync.sync_client import SyncClient
 from syftbox.lib.constants import PERM_FILE
 from syftbox.lib.workspace import SyftWorkspace
+from syftbox.server.migrations import run_migrations
 from syftbox.server.server import app
 from syftbox.server.settings import ServerSettings
 
@@ -44,9 +45,11 @@ def client(monkeypatch, tmp_path):
     """Every client gets their own snapshot folder at `tmp_path`"""
     snapshot_folder = tmp_path / "snapshot"
     settings = ServerSettings.from_data_folder(snapshot_folder)
+
     monkeypatch.setenv("SYFTBOX_DATA_FOLDER", str(settings.data_folder))
     monkeypatch.setenv("SYFTBOX_SNAPSHOT_FOLDER", str(settings.snapshot_folder))
     monkeypatch.setenv("SYFTBOX_USER_FILE_PATH", str(settings.user_file_path))
+    monkeypatch.setenv("SYFTBOX_OTEL_ENABLED", str(False))
 
     datasite_name = TEST_DATASITE_NAME
     datasite = settings.snapshot_folder / datasite_name
@@ -66,6 +69,7 @@ def client(monkeypatch, tmp_path):
     permfile.touch()
     permfile.write_text(json.dumps(PERMFILE_DICT))
 
+    run_migrations(settings)
     with TestClient(app) as client:
         access_token = get_access_token(client, TEST_DATASITE_NAME)
         client.headers["Authorization"] = f"Bearer {access_token}"
@@ -88,9 +92,12 @@ def sync_client(client: TestClient, tmp_path: Path):
 def client_without_perms(monkeypatch, tmp_path):
     """Every client gets their own snapshot folder at `tmp_path`"""
     settings = ServerSettings.from_data_folder(tmp_path)
+    settings.otel_enabled = False
+
     monkeypatch.setenv("SYFTBOX_DATA_FOLDER", str(settings.data_folder))
     monkeypatch.setenv("SYFTBOX_SNAPSHOT_FOLDER", str(settings.snapshot_folder))
     monkeypatch.setenv("SYFTBOX_USER_FILE_PATH", str(settings.user_file_path))
+    monkeypatch.setenv("SYFTBOX_OTEL_ENABLED", str(False))
 
     datasite_name = TEST_DATASITE_NAME
     datasite = settings.snapshot_folder / datasite_name
@@ -104,6 +111,7 @@ def client_without_perms(monkeypatch, tmp_path):
     permfile.touch()
     permfile.write_text("")
 
+    run_migrations(settings)
     with TestClient(app) as client:
         access_token = get_access_token(client, TEST_DATASITE_NAME)
         client.headers["Authorization"] = f"Bearer {access_token}"
